Create Attributes.cs to separate concerns between reflectively extracting attributes and utilites for determining types in Types.cs
